+++ react/components/EditorContainer/Topbar/ContextSelectors/LocaleSelector.tsx
@@ -20,13 +20,44 @@ const LocaleSelector: React.FC<Props> = ({
 }) => {
   const { culture, emitter } = iframeRuntime
 
+  const editor = useEditorContext()
+
   const [locale, setLocale] = React.useState(culture.locale)
 
   const handleChange = React.useCallback(
     (e: React.ChangeEvent<HTMLSelectElement>) => {
-      setLocale(e.target.value)
+      editor.setMode('disabled')
 
-      emitter.emit('localesChanged', e.target.value)
+      setLocale(e?.target?.value)
+
+      emitter.emit(
+        'localesChanged',
+        e?.target?.value,
+        null,
+        (locale: string) => {
+          let bindingQueryString
+
+          if (editor.iframeWindow) {
+            const searchParams = new URLSearchParams(
+              editor?.iframeWindow?.location.search
+            )
+
+            const bindingAddress = searchParams.get('__bindingAddress')
+
+            if (bindingAddress) {
+              bindingQueryString = `__bindingAddress=${decodeURIComponent(
+                bindingAddress
+              )}`
+            }
+
+            if (e?.target?.value) {
+              editor.iframeWindow.location.search = `${bindingQueryString}&__locale=${e.target.value}&__siteEditor=true`
+            } else {
+              editor.iframeWindow.location.search = `${bindingQueryString}&__locale=${locale}&__siteEditor=true`
+            }
+          }
+        }
+      )
     },
     [emitter]
   )
@@ -37,6 +68,12 @@ const LocaleSelector: React.FC<Props> = ({
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [culture.locale])
+
+  React.useEffect(() => {
+    if (editor.mode !== 'content') {
+      editor.setMode('content')
+    }
+  }, [culture.locale, culture.country])
 
   const Selector = React.useCallback(
     () => (
@@ -53,7 +90,6 @@ const LocaleSelector: React.FC<Props> = ({
     [className, handleChange, isDisabled, locale, options]
   )
 
-  const editor = useEditorContext()
   const intl = useIntl()
 
   return editor.editTreePath ? (